+++ repository/src/main/java/com/epam/esm/models/orders/OrderModel.java
@@ -4,7 +4,6 @@ import com.epam.esm.models.certificates.OnlyCertificate;
 import com.epam.esm.models.users.ReadUserModel;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -18,6 +17,6 @@ public class OrderModel {
     private long id;
     ReadUserModel userModel;
     List<OnlyCertificate> certificates;
-    private BigDecimal cost;
+    private double cost;
     private LocalDateTime datePurchase;
 }
